Add wired and selection packages to organize features
