+++ audioswitch/src/test/java/com/twilio/audioswitch/selection/AudioDeviceSelectorTest.kt
@@ -1,4 +1,4 @@
-package com.twilio.audioswitch
+package com.twilio.audioswitch.selection
 
 import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothDevice
@@ -16,17 +16,18 @@ import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyZeroInteractions
 import com.nhaarman.mockitokotlin2.whenever
-import com.twilio.audioswitch.AudioDevice.Type.EARPIECE
-import com.twilio.audioswitch.AudioDevice.Type.SPEAKERPHONE
-import com.twilio.audioswitch.AudioDeviceSelector.State.ACTIVATED
-import com.twilio.audioswitch.AudioDeviceSelector.State.STARTED
-import com.twilio.audioswitch.AudioDeviceSelector.State.STOPPED
+import com.twilio.audioswitch.selection.AudioDevice.Type.EARPIECE
+import com.twilio.audioswitch.selection.AudioDevice.Type.SPEAKERPHONE
+import com.twilio.audioswitch.selection.AudioDeviceSelector.State.ACTIVATED
+import com.twilio.audioswitch.selection.AudioDeviceSelector.State.STARTED
+import com.twilio.audioswitch.selection.AudioDeviceSelector.State.STOPPED
 import com.twilio.audioswitch.android.BuildWrapper
 import com.twilio.audioswitch.android.LogWrapper
 import com.twilio.audioswitch.bluetooth.BluetoothController
 import com.twilio.audioswitch.bluetooth.BluetoothControllerAssertions
 import com.twilio.audioswitch.bluetooth.BluetoothHeadsetReceiver
 import com.twilio.audioswitch.bluetooth.PreConnectedDeviceListener
+import com.twilio.audioswitch.wired.WiredHeadsetReceiver
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.CoreMatchers.nullValue
